engines/weaver/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.0.26
langchain>=0.1.12
pydantic>=2.6.4
python-dotenv>=1.0.1
httpx>=0.27.0
# For module-to-module communication
